% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnanalysis.R
\name{gsanalysis}
\alias{gsanalysis}
\title{Grow-Shrink Analysis}
\usage{
gsanalysis(x)
}
\arguments{
\item{data}{set}
}
\value{
model summary, undirected arcs, model structure as plot
}
\description{
This function applies the grow-shrink algorithm to a given data set and
returns missing node connections and a model summary
}
