% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnanalysis.R
\name{hcanalysis}
\alias{hcanalysis}
\title{Hill-Climbing Analysis}
\usage{
hcanalysis(x)
}
\arguments{
\item{data}{set}
}
\value{
model summary, undirected arcs, model structure as plot
}
\description{
This function applies the hill-climbing algorithm to a given data set and
returns missing node connections and a model summary
}
